refactor(github-analysis): split BonusMarksTab into smaller components
- Created new hooks: useBonusMarks and useContributors
- Moved styles and types into separate files
- Simplified the main BonusMarksTab component
